+++ examples/main/modules.tf
@@ -28,8 +28,10 @@ module "storage_sas_token" {
   source  = "claranet/storage-sas-token/azurerm"
   version = "x.x.x"
 
-  storage_account_name = azurerm_storage_account.my_storage.name
-  resource_group_name  = module.rg.resource_group_name
+  storage_account_name              = azurerm_storage_account.my_storage.name
+  storage_account_connection_string = azurerm_storage_account.my_storage.primary_connection_string
+
+  resource_group_name = module.rg.resource_group_name
 }
 
 ### Service SAS for a container
@@ -43,7 +45,10 @@ module "container_sas_token" {
   source  = "claranet/storage-sas-token/azurerm"
   version = "x.x.x"
 
-  storage_account_name = azurerm_storage_account.my_storage.name
-  resource_group_name  = module.rg.resource_group_name
-  storage_container    = azurerm_storage_container.my_container.name
+  storage_account_name                  = azurerm_storage_account.my_storage.name
+  storage_account_connection_string     = azurerm_storage_account.my_storage.primary_connection_string
+  storage_account_primary_blob_endpoint = azurerm_storage_account.my_storage.primary_blob_endpoint
+
+  resource_group_name = module.rg.resource_group_name
+  storage_container   = azurerm_storage_container.my_container.name
 }