+++ r-sas-token.tf
@@ -1,18 +1,10 @@
-data "azurerm_storage_account" "storage" {
-  count = var.enabled ? 1 : 0
-
-  name                = var.storage_account_name
-  resource_group_name = var.resource_group_name
-}
-
 data "external" "generate_storage_sas_token" {
-  count = var.enabled ? 1 : 0
 
   program = ["bash", "${path.module}/files/script_sas_token.sh"]
 
   query = {
-    storage_account_name      = join("", data.azurerm_storage_account.storage.*.name)
-    storage_connection_string = join("", data.azurerm_storage_account.storage.*.primary_connection_string)
+    storage_account_name      = local.sa_name
+    storage_connection_string = var.storage_account_connection_string
     storage_container         = var.storage_container
     token_expiry              = var.sas_token_expiry
     services                  = var.services
@@ -21,3 +13,5 @@ data "external" "generate_storage_sas_token" {
     permissions_container     = var.permissions_container
   }
 }
+
+