+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased 
 
 Fixed 
-  * AZ-674 : fix this bug  https://claranet.slack.com/archives/CGE40D3KL/p1643202262270200
+  * AZ-674 : fix sas token generation bug (datasource that reference a non created resource). Clean unused module variables.
 
 # v4.1.0 - 2022-01-12